+++ aya-log/aya-log/Cargo.toml
@@ -19,5 +19,9 @@ log = "0.4"
 bytes = "1.1"
 tokio = { version = "1.2.0" }
 
+[dev-dependencies]
+simplelog = "0.12"
+testing_logger = "0.1.1"
+
 [lib]
 path = "src/lib.rs"
